P3-E1

+++ matrix.h
@@ -20,7 +20,7 @@ t_matrix create_adj_matrix(t_adj_list);
 void free_matrix(t_matrix);
 t_matrix copy_matrix(t_matrix);
 t_matrix mult_matrix(t_matrix, t_matrix);
-int diff_matrix(t_matrix, t_matrix);
+float diff_matrix(t_matrix, t_matrix);
 void print_matrix(t_matrix);
 
 /**
